+++ src/testing.ts
@@ -2,7 +2,13 @@
 export { ipv4ToBuffer } from "./utils/ip";
 export { encodeEasyMPLength } from "./easymp/length";
 export { createEprdPacket, createTile } from "./easymp/packet";
-export { createConnectPacket, createControlPacket } from "./easymp/session";
+export {
+  createConnectPacket,
+  createControlPacket,
+  createDiscoveryPacket,
+  createMovieStartPacket,
+} from "./easymp/session";
+export { createEasyMPFfmpegArgs } from "./easymp/movie";
 export { parseEscVpResponse } from "./control/escvp";
 export { powerStateFromCode } from "./projector";
 export { sameSubnet } from "./utils/network";